Hide warning "narrowing conversion of ‘x1’ from ‘int’ to ‘uint16_t {aka short unsigned int}’ inside { } is ill-formed in C++11"

+++ osd.c
@@ -544,7 +544,7 @@ void cXinelibOsd::Flush(void)
       int NumColors;
       const tColor *Colors = Bitmap->Colors(NumColors);
       if (Colors) {
-	osd_rect_t DirtyArea = {x1:x1, y1:y1, x2:x2, y2:y2};
+	osd_rect_t DirtyArea = {x1:(uint16_t)x1, y1:(uint16_t)y1, x2:(uint16_t)x2, y2:(uint16_t)y2};
         CmdLut8(i,
                 Left() + Bitmap->X0() + XOffset, Top() + Bitmap->Y0() + YOffset,
                 Bitmap->Width(), Bitmap->Height(),
